refactor: files structure for search and related
