+++ client-app/core/constants/environment.ts
@@ -1,0 +1,1 @@
+export const IS_DEVELOPMENT = import.meta.env.MODE === "development";
