+++ client-app/core/constants/index.ts
@@ -1,0 +1,6 @@
+export * from "./environment";
+export * from "./notifications";
+export * from "./products";
+export * from "./search";
+export * from "./security";
+export * from "./wishlists";
