+++ client-app/core/constants/notifications.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_NOTIFICATION_DURATION = 15000;
