+++ client-app/core/utilities/index.ts
@@ -2,8 +2,8 @@ export * from "./address";
 export * from "./breadcrumbs";
 export * from "./categories";
 export * from "./common";
-export * from "./facets";
 export * from "./line-items";
 export * from "./logger";
 export * from "./menu";
 export * from "./product";
+export * from "./search";
