+++ client-app/core/utilities/search/index.ts
@@ -1,0 +1,2 @@
+export * from "./facets";
+export * from "./sorting";
